Thread
moved getAddress to different thread
